backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Brandon Lamer-Connolly", email = "brandon.lc.1334@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-gis (>=1.2.0,<2.0.0)",
    "dj-database-url (>=2.1.0,<3.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "timezonefinder (>=8.2.1,<9.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "gunicorn (>=23.0.0)",
    "whitenoise (>=6.8.2)",
    "django-phonenumber-field[phonenumbers] (>=8.4.0,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false